+++ src/elements/gunModifier.ts
@@ -1,49 +1,35 @@
 import * as mat from "transformation-matrix";
 
 import { Gun } from "../gun";
-import { FiringState, FireDataModifier, FireData } from "../firing-state";
+import { FiringState, FireData } from "../firing-state";
 import {
   TConstantOrLazy,
-  calcValueFromConstantOrLazy,
   calcTransFormFromConstantOrLazy
 } from "../lazyEvaluative";
-import { VirtualMuzzleGenerator } from "../muzzle";
 import { decomposeTransform } from "../transform-util";
 
 /**
- * ModifierGun update FireData when fired or immediately.
+ * ModifierGun update FireData when fired.
  */
 export class ModifierGun implements Gun {
-  private readonly modifyLater: boolean;
   private readonly modifier: FireDataModifier;
+
   /**
-   * @param modifyLater Use modifier later or immediately
    * @param modifier Used modifier
    */
-  public constructor(modifyLater: boolean, modifier: FireDataModifier) {
-    this.modifyLater = modifyLater;
+  public constructor(modifier: FireDataModifier) {
     this.modifier = modifier;
   }
 
   public *play(state: FiringState): IterableIterator<void> {
-    ModifierGun.modifyImmediatelyOrLater(
-      state,
-      this.modifyLater,
-      this.modifier
-    );
+    state.pushModifier(this.modifier.createModifier(state));
   }
+}
 
-  private static modifyImmediatelyOrLater(
-    state: FiringState,
-    modifyLater: boolean,
-    modifier: FireDataModifier
-  ): void {
-    if (modifyLater) {
-      state.pushModifier(modifier);
-    } else {
-      modifier.modifyFireData(state, state.fireData);
-    }
-  }
+export interface FireDataModifier {
+  createModifier(
+    state: FiringState
+  ): (stateConst: FiringState, fireData: FireData) => void;
 }
 
 /**
@@ -55,9 +41,14 @@ export class TransformModifier implements FireDataModifier {
     this.trans = trans;
   }
 
-  public modifyFireData(stateConst: FiringState, fireData: FireData): void {
-    const transConst = calcTransFormFromConstantOrLazy(stateConst, this.trans);
-    fireData.transform = mat.transform(transConst, fireData.transform);
+  public createModifier(
+    state: FiringState
+  ): (stateConst: FiringState, fireData: FireData) => void {
+    const transConst = calcTransFormFromConstantOrLazy(state, this.trans);
+
+    return (_stateConst: FiringState, fireData: FireData): void => {
+      fireData.transform = mat.transform(transConst, fireData.transform);
+    };
   }
 }
 
@@ -82,38 +73,23 @@ export class InvertTransformModifier implements FireDataModifier {
     this.option = option;
   }
 
-  public modifyFireData(stateConst: FiringState, fireData: FireData): void {
-    const [t, angleDeg, scale] = decomposeTransform(fireData.transform);
+  public createModifier(
+    _state: FiringState
+  ): (stateConst: FiringState, fireData: FireData) => void {
     const xRate = this.option.translationX ? -1 : 1;
     const yRate = this.option.translationY ? -1 : 1;
     const angleRate = this.option.angle ? -1 : 1;
-    const translateNew = { x: t.x * xRate, y: t.y * yRate };
-    const angleDegNew = angleDeg * angleRate;
-    fireData.transform = mat.transform(
-      mat.translate(translateNew.x, translateNew.y),
-      mat.rotateDEG(angleDegNew),
-      mat.scale(scale.x, scale.y)
-    );
-  }
-}
 
-/**
- * Set parameter in FireData when played.
- */
-export class SetParameterImmediatelyModifier implements FireDataModifier {
-  private readonly name: string;
-  private readonly value: TConstantOrLazy<number>;
-
-  public constructor(name: string, value: TConstantOrLazy<number>) {
-    this.name = name;
-    this.value = value;
-  }
-
-  public modifyFireData(stateConst: FiringState, fireData: FireData): void {
-    fireData.parameters.set(
-      this.name,
-      calcValueFromConstantOrLazy(stateConst, this.value)
-    );
+    return (_stateConst: FiringState, fireData: FireData): void => {
+      const [t, angleDeg, scale] = decomposeTransform(fireData.transform);
+      const translateNew = { x: t.x * xRate, y: t.y * yRate };
+      const angleDegNew = angleDeg * angleRate;
+      fireData.transform = mat.transform(
+        mat.translate(translateNew.x, translateNew.y),
+        mat.rotateDEG(angleDegNew),
+        mat.scale(scale.x, scale.y)
+      );
+    };
   }
 }
 
@@ -135,66 +111,14 @@ export class ModifyParameterModifier implements FireDataModifier {
     this.modifier = modifier;
   }
 
-  public modifyFireData(stateConst: FiringState, fireData: FireData): void {
-    const oldValue = fireData.parameters.get(this.name);
-    if (oldValue === undefined)
-      throw new Error(`parameter <${this.name}> was not set`);
-    fireData.parameters.set(this.name, this.modifier(stateConst, oldValue));
-  }
-}
-
-/**
- * Set text in FireData when played.
- */
-export class SetTextImmediatelyModifier implements FireDataModifier {
-  private readonly name: string;
-  private readonly text: TConstantOrLazy<string>;
-
-  public constructor(name: string, text: TConstantOrLazy<string>) {
-    this.name = name;
-    this.text = text;
-  }
-
-  public modifyFireData(stateConst: FiringState, fireData: FireData): void {
-    fireData.texts.set(
-      this.name,
-      calcValueFromConstantOrLazy(stateConst, this.text)
-    );
-  }
-}
-
-/**
- * Set muzzle in FireData when played.
- */
-export class SetMuzzleImmediatelyModifier implements FireDataModifier {
-  private readonly name: TConstantOrLazy<string>;
-
-  public constructor(name: TConstantOrLazy<string>) {
-    this.name = name;
-  }
-
-  public modifyFireData(stateConst: FiringState, _fireData: FireData): void {
-    const muzzleName = calcValueFromConstantOrLazy(stateConst, this.name);
-    stateConst.muzzle = stateConst.getMuzzleByName(muzzleName);
-  }
-}
-
-/**
- * Attach virtual muzzle to current muzzle.
- */
-export class AttachVirtualMuzzleImmediatelyModifier
-  implements FireDataModifier {
-  private readonly virtualMuzzleGenerator: VirtualMuzzleGenerator;
-
-  public constructor(virtualMuzzleGenerator: VirtualMuzzleGenerator) {
-    this.virtualMuzzleGenerator = virtualMuzzleGenerator;
-  }
-
-  public modifyFireData(stateConst: FiringState, _fireData: FireData): void {
-    if (stateConst.muzzle === null)
-      throw new Error("Muzzle was not set at FiringState");
-    const muzzle = this.virtualMuzzleGenerator.generate();
-    muzzle.basedOn(stateConst.muzzle);
-    stateConst.muzzle = muzzle;
+  public createModifier(
+    _state: FiringState
+  ): (stateConst: FiringState, fireData: FireData) => void {
+    return (stateConst: FiringState, fireData: FireData): void => {
+      const oldValue = fireData.parameters.get(this.name);
+      if (oldValue === undefined)
+        throw new Error(`parameter <${this.name}> was not set`);
+      fireData.parameters.set(this.name, this.modifier(stateConst, oldValue));
+    };
   }
 }