+++ src/firing-state.ts
@@ -22,7 +22,9 @@ export interface FiringState {
    *
    * @param modifier modifier would applied when fire bullet
    */
-  pushModifier(modifier: FireDataModifier): void;
+  pushModifier(
+    modifier: (stateConst: FiringState, fireData: FireData) => void
+  ): void;
 
   /** Calculate modified fire data. */
   calcModifiedFireData(): FireData;
@@ -61,13 +63,6 @@ export interface FireData {
 
   /** Copy this data. */
   copy(): FireData;
-}
-
-/**
- * IFireDataModifier modify FireData.
- */
-export interface FireDataModifier {
-  modifyFireData(stateConst: FiringState, fireData: FireData): void;
 }
 
 /**
@@ -119,7 +114,10 @@ export class DefaultFiringState implements FiringState {
   public muzzle: Muzzle | null;
 
   /** Function would applied to fireData when fire bullet, */
-  private readonly modifiers: FireDataModifier[];
+  private readonly modifiers: ((
+    stateConst: FiringState,
+    fireData: FireData
+  ) => void)[];
 
   /** Player playing with this state */
   private readonly player: Player;
@@ -152,7 +150,9 @@ export class DefaultFiringState implements FiringState {
    *
    * @param modifier Function would applied when fire bullet
    */
-  public pushModifier(modifier: FireDataModifier): void {
+  public pushModifier(
+    modifier: (stateConst: FiringState, fireData: FireData) => void
+  ): void {
     this.modifiers.push(modifier);
   }
 
@@ -162,9 +162,10 @@ export class DefaultFiringState implements FiringState {
     const fdClone = this.fireData.copy();
 
     // Apply modifiers
-    this.modifiers
-      .reverse()
-      .map((mod): void => mod.modifyFireData(this, fdClone));
+    const reversedMods = [...this.modifiers].reverse();
+    for (const mod of reversedMods) {
+      mod(this, fdClone);
+    }
 
     // Apply muzzle transform
     fdClone.transform = mat.transform(