+++ CHANGELOG.md
@@ -7,7 +7,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update ReadMe
 
-## [0.6.2] - 2019-07-17
+## [0.6.6] - 2019-12-01
+### Fixed
+- Fix angle inverting (mirror, alternate and invert)
+
+## [0.6.5] - 2019-11-29
+### Fixed
+- Fix guns exporting
+
+## [0.6.4] - 2019-11-29
+### Fixed
+- Fix parameter changing
+
+## [0.6.3] - 2019-07-17
 ### Changed
 - Update dependencies
 
@@ -98,7 +110,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/sankaku-deltalab/guntree/compare/0.6.3...HEAD
+[Unreleased]: https://github.com/sankaku-deltalab/guntree/compare/0.6.6...HEAD
+[0.6.6]: https://github.com/sankaku-deltalab/guntree/compare/0.6.5...0.6.6
+[0.6.5]: https://github.com/sankaku-deltalab/guntree/compare/0.6.4...0.6.5
+[0.6.4]: https://github.com/sankaku-deltalab/guntree/compare/0.6.3...0.6.4
 [0.6.3]: https://github.com/sankaku-deltalab/guntree/compare/0.6.2...0.6.3
 [0.6.2]: https://github.com/sankaku-deltalab/guntree/compare/0.6.1...0.6.2
 [0.6.1]: https://github.com/sankaku-deltalab/guntree/compare/0.6.0...0.6.1